moved breed classifier into /api
